% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{create_accessors_s3}
\alias{create_accessors_s3}
\title{Create table accessors for S3}
\usage{
create_accessors_s3(date = "latest", env)
}
\arguments{
\item{date}{The date for the tables to be accessed}

\item{env}{The environment into which the accessor functions should be
written}
}
\description{
Create table accessors for S3
}
