% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{create_s3_connector}
\alias{create_s3_connector}
\title{Create connection to Amazon S3}
\usage{
create_s3_connector(
  access_key_id = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret_access_key = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  region = Sys.getenv("AWS_DEFAULT_REGION")
)
}
\arguments{
\item{access_key_id}{String containing the S3 access key.}

\item{secret_access_key}{String containing the S3 secret access key.}

\item{region}{String containing the S3 default region.}
}
\description{
Creates a connector to the Amazon Simple Storage Service where DataCamp Data
Connector data is hosted.
}
\note{
Find the arguments to this function in the Admin Portal for
your DataCamp group. See Reporting -> Export -> Data Connector -> View
Configuration Details. It is recommended to store these values as environment
variables.
}
\examples{
\dontrun{
Sys.setenv(
  AWS_ACCESS_KEY_ID = "your access key ID",
  AWS_SECRET_ACCESS_KEY = "your secret key",
  AWS_DEFAULT_REGION = "your default region"
)
}
s3_conn <- create_s3_connector()
str(s3_conn, 1)
}
\seealso{
\code{\link[paws]{s3}}
}
