% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\docType{data}
\name{DC_DATASETS}
\alias{DC_DATASETS}
\title{Dataset names}
\format{
An object of class \code{character} of length 15.
}
\usage{
DC_DATASETS
}
\value{
Character vector of names of datasets.
}
\description{
Names of datasets available via data connector.
}
\examples{
DC_DATASETS
}
\keyword{datasets}
