% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcdcr-package.R
\docType{package}
\name{dcdcr-package}
\alias{dcdcr}
\alias{dcdcr-package}
\title{dcdcr: DataCamp Data Connector utilities in R}
\description{
Utilities to make it easy to import and manipulate data from 
  DataCamp Data Connector.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datacamp/dcdcr}
  \item \url{https://datacamp.github.io/dcdcr}
  \item Report bugs at \url{https://github.com/datacamp/dcdcr/issues}
}

}
\author{
\strong{Maintainer}: Richard Cotton \email{richie@datacamp.com}

Authors:
\itemize{
  \item Ramnath Vaidyanathan \email{ramnath@datacamp.com}
}

}
\keyword{internal}
