% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get_dc_datasets}
\alias{get_dc_datasets}
\title{Get DataCamp Data Connector datasets}
\usage{
get_dc_datasets(
  s3_sess,
  datasets = DC_DATASETS,
  date = "latest",
  bucket = Sys.getenv("AWS_S3_BUCKET_NAME"),
  verbose = getOption("verbose")
)
}
\arguments{
\item{s3_sess}{An S3 data connector, as returned by \code{create_s3_sessector}.}

\item{datasets}{Names of the datasets to import.}

\item{date}{A Date or string in \code{\%Y-\%m-\%d} format denoting the date at
which to retrieve data for. Use \code{"latest"} to get the most recent data.}

\item{bucket}{A string containing the DataCamp Data Connector bucket name.}

\item{verbose}{If \code{TRUE}, print a message of the date of the data.}
}
\value{
A list of tibbles containing the datasets. The date is stored in an
attribute named \code{date}
}
\description{
Gets DataCamp Data Connector datasets from S3.
}
\examples{
\dontrun{
Sys.setenv(
  AWS_ACCESS_KEY_ID = "your access key ID",
  AWS_SECRET_ACCESS_KEY = "your secret key",
  AWS_DEFAULT_REGION = "your default region",
  AWS_S3_BUCKET_NAME = "your bucket"
)
}
s3_sess <- create_s3_sessector()
dc <- get_dc_datasets(s3_sess)
str(dc, 1, give.attr = FALSE)
}
\seealso{
\code{\link{create_s3_sessector}}, \code{\link{DC_DATASETS}}
}
