% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get_most_recent_date}
\alias{get_most_recent_date}
\title{Get the most recent date where data is available}
\usage{
get_most_recent_date(
  s3_sess,
  bucket = Sys.getenv("AWS_S3_BUCKET_NAME"),
  verbose = getOption("verbose")
)
}
\arguments{
\item{s3_sess}{An S3 data connector, as returned by \code{create_s3_sessector}.}

\item{bucket}{A string containing the DataCamp Data Connector bucket name.}

\item{verbose}{If \code{TRUE}, print a message of the date of the data.}
}
\value{
A string containing the date in \code{\%Y-\%m-\%d} format.
}
\description{
Get the most recent date where DataCamp Data Connector data is available.
}
\note{
Find the bucket argument to this function in the Admin Portal for
your DataCamp group. See Reporting -> Export -> Data Connector -> View
Configuration Details. It is recommended to store this value as an
environment variable.
}
\examples{
\dontrun{
Sys.setenv(
  AWS_ACCESS_KEY_ID = "your access key ID",
  AWS_SECRET_ACCESS_KEY = "your secret key",
  AWS_DEFAULT_REGION = "your default region",
  AWS_S3_BUCKET_NAME = "your bucket"
)
}
s3_sess <- create_s3_sessector()
get_most_recent_date(s3_sess, verbose = TRUE)
}
\seealso{
\code{\link{create_s3_sessector}}
}
