% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{list_objects_s3}
\alias{list_objects_s3}
\title{List objects in an S3 bucket}
\usage{
list_objects_s3(bucket = Sys.getenv("AWS_S3_BUCKET_NAME"), date = "latest")
}
\arguments{
\item{bucket}{The name of the S3 bucket.}

\item{date}{The date to filter the objects on.}
}
\description{
List objects in an S3 bucket
}
\examples{
\dontrun{
list_objects_s3() \%>\%
  head()
}
}
\keyword{internal}
