% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_tbl}
\alias{s3_tbl}
\title{Read table from S3}
\usage{
s3_tbl(x, date = "latest")
}
\arguments{
\item{x}{The name of the table}

\item{date}{The date when it was generated (defaults to "latest")}
}
\description{
Read table from S3
}
\examples{
\dontrun{
tbl_s3('chapter_dim')
}
}
