% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{create_s3_session}
\alias{create_s3_session}
\title{Create Amazon S3 session}
\usage{
create_s3_session(
  access_key_id = get_env_var_aws_access_key(),
  secret_access_key = get_env_var_aws_secret(),
  region = get_env_var_aws_region()
)
}
\arguments{
\item{access_key_id}{String containing the S3 access key.}

\item{secret_access_key}{String containing the S3 secret access key.}

\item{region}{String containing the S3 default region.}
}
\description{
Creates an Amazon Simple Storage Service session. (This is where DataCamp
Data Connector data is hosted).
}
\note{
Find the arguments to this function in the Admin Portal for
your DataCamp group. See Reporting -> Export -> Data Connector -> View
Configuration Details. It is recommended to store these values as environment
variables.
}
\examples{
\dontrun{
Sys.setenv(
  AWS_ACCESS_KEY_ID = "your access key ID",
  AWS_SECRET_ACCESS_KEY = "your secret key",
  AWS_REGION = "your region"
)
}
s3_sess <- create_s3_session()
str(s3_sess, 1)
}
\seealso{
\code{\link[paws]{s3}}
}
